requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
langchain>=0.0.350
langchain-community>=0.0.1
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.11  # Optional, for voice input
